requirements.txt
flask==1.1.4
google.generativeai
markupsafe==1.1.1
